chore: Refactor PublishReporter creation
